+++ doc/changelog.rst
@@ -62,6 +62,7 @@ Addons
 Front
 ~~~~~
 
+- Cache template helpers, sorts and filters using context cache
 - Enable password reset when shop is in maintenance mode
 - Shop can now have a favicon
 - Variation children that are not purchaseable should not be visible anymore in dropdowns